2927: Refactored poster single component
